Move MainWindow and MainViewModel code to TabContentView and TabContentViewModel
