+++ WebBrowser.WPF/Views/TabContentView.xaml
@@ -1,0 +1,116 @@
+﻿<reactiveUi:ReactiveUserControl x:TypeArguments="viewModels:TabContentViewModel" 
+                                x:Class="WebBrowser.WPF.Views.TabContentView"
+                                xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+                                xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+                                xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+                                xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+                                xmlns:local="clr-namespace:WebBrowser.WPF.Views"
+                                xmlns:reactiveUi="http://reactiveui.net"
+                                xmlns:controls="clr-namespace:WebBrowser.WPF.Controls"
+                                PreviewKeyDown="OnUserControl_PreviewKeyDown"
+                                xmlns:converters="clr-namespace:WebBrowser.WPF.Converters"
+                                xmlns:wpf="clr-namespace:CefSharp.Wpf;assembly=CefSharp.Wpf"
+                                xmlns:viewModels="clr-namespace:WebBrowser.ViewModels;assembly=WebBrowser"
+                                mc:Ignorable="d"
+                                d:DesignHeight="300" d:DesignWidth="300">
+    <Grid>
+        <Grid.ColumnDefinitions>
+            <ColumnDefinition Width="Auto" />
+            <ColumnDefinition Width="Auto" />
+            <ColumnDefinition Width="Auto" />
+            <ColumnDefinition Width="*" />
+            <ColumnDefinition Width="Auto" />
+        </Grid.ColumnDefinitions>
+        <Grid.RowDefinitions>
+            <RowDefinition Height="50" />
+            <RowDefinition Height="*" />
+        </Grid.RowDefinitions>
+        <Button Grid.Column="0"
+                Style="{StaticResource IconButtons}"
+                Command="{Binding BackCommand, ElementName=BrowserPane}">
+            <Button.ToolTip>
+                <ToolTip Content="Click to go back" />
+            </Button.ToolTip>
+            <Image Source="{StaticResource ArrowLeft}" />
+        </Button>
+        <Button Grid.Column="1"
+                Style="{StaticResource IconButtons}"
+                Command="{Binding ForwardCommand, ElementName=BrowserPane}">
+            <Button.ToolTip>
+                <ToolTip Content="Click to go forward" />
+            </Button.ToolTip>
+            <Image Source="{StaticResource ArrowRight}" />
+        </Button>
+        <Button Grid.Column="2"
+                Style="{StaticResource IconButtons}"
+                IsEnabled="{Binding Title, ElementName=BrowserPane, Converter={x:Static converters:ObjectToBooleanConverter.Instance}}"
+                Command="{Binding ReloadCommand, ElementName=BrowserPane}">
+            <Button.ToolTip>
+                <ToolTip Content="Click to go reload this page" />
+            </Button.ToolTip>
+            <Image Source="{StaticResource Reload}" />
+        </Button>
+        <TextBox x:Name="AddressBarTextBox"
+                 Grid.Column="3"
+                 PreviewKeyDown="OnTextBox_PreviewKeyDown"
+                 Text="{Binding AddressBarText, UpdateSourceTrigger=PropertyChanged}">
+            <TextBox.InputBindings>
+                <KeyBinding Command="{Binding Go}"
+                            CommandParameter="{Binding Path=Text, RelativeSource={RelativeSource AncestorType=TextBox}}"
+                            Key="Enter">
+                </KeyBinding>
+            </TextBox.InputBindings>
+        </TextBox>
+        <Popup
+            Margin="0,-5,0,0"
+            IsOpen="{Binding Suggestions.Count, Converter={x:Static converters:ObjectToBooleanConverter.Instance}, Mode=OneWay}"
+            PlacementTarget="{Binding ElementName=AddressBarTextBox}"
+            Placement="Bottom"
+            Width="{Binding Path=ActualWidth, ElementName=AddressBarTextBox}"
+            StaysOpen="False">
+            <ListBox x:Name="SuggestionsListBox" 
+                     ItemsSource="{Binding Suggestions}"
+                     SelectedItem="{Binding SelectedSuggestion, Mode=OneWayToSource, UpdateSourceTrigger=PropertyChanged}"
+                     ScrollViewer.HorizontalScrollBarVisibility="Disabled">
+                <ListBox.ItemContainerStyle>
+                    <Style TargetType="ListBoxItem" BasedOn="{StaticResource {x:Type ListBoxItem}}">
+                        <EventSetter Event="PreviewMouseLeftButtonDown" Handler="ListBoxItem_OnPreviewMouseLeftButtonDown" />
+                    </Style>
+                </ListBox.ItemContainerStyle>
+                <ListBox.ItemTemplate>
+                    <DataTemplate>
+                        <TextBlock Text="{Binding }"
+                                   TextTrimming="CharacterEllipsis" />
+                    </DataTemplate>
+                </ListBox.ItemTemplate>
+            </ListBox>
+        </Popup>
+        <Button Grid.Column="4"
+                Style="{StaticResource IconButtons}"
+                Command="{Binding ClearHistory}">
+            <Button.ToolTip>
+                <ToolTip Content="Click to clear search history" />
+            </Button.ToolTip>
+            <Image Source="{StaticResource Trash}" />
+        </Button>
+
+        <ScrollViewer Grid.Row="1" Grid.ColumnSpan="5"
+                      HorizontalScrollBarVisibility="Disabled"
+                      VerticalScrollBarVisibility="Hidden">
+            <StackPanel>
+                <local:WelcomeView  Visibility="{Binding BrowserAddress, Converter={x:Static converters:ObjectToVisibilityConverter.Instance}, ConverterParameter='!'}" />
+                <local:BrowserHistoryView DataContext="{Binding BrowserHistory}"
+                                          Visibility="{Binding DataContext.BrowserAddress, RelativeSource={RelativeSource AncestorType=Window} , Converter={x:Static converters:ObjectToVisibilityConverter.Instance}, ConverterParameter='!'}" />
+            </StackPanel>
+        </ScrollViewer>
+        <wpf:ChromiumWebBrowser x:Name="BrowserPane"
+                                Grid.Row="1"
+                                Grid.ColumnSpan="5"
+                                Visibility="{Binding BrowserAddress, Converter={x:Static converters:ObjectToVisibilityConverter.Instance}}"
+                                Address="{Binding BrowserAddress, Mode=TwoWay}">
+            <wpf:ChromiumWebBrowser.MenuHandler>
+                <controls:ContextMenuHandler />
+            </wpf:ChromiumWebBrowser.MenuHandler>
+        </wpf:ChromiumWebBrowser>
+    </Grid>
+</reactiveUi:ReactiveUserControl>
