+++ WebBrowser.WPF/WebBrowser.WPF.csproj
@@ -14,6 +14,7 @@
 	</PropertyGroup>
 	
 	<ItemGroup>
+		<ProjectReference Include="..\WebBrowser.Test\WebBrowser.Test.csproj" />
 		<ProjectReference Include="..\WebBrowser\WebBrowser.csproj" />
 	</ItemGroup>
